SCMOD-4658: Refactor Application.java to call the class from container startup
