+++ util-liquibase-installer/src/main/java/com/hpe/caf/utilliquibaseinstaller/Application.java
@@ -15,16 +15,8 @@
  */
 package com.hpe.caf.utilliquibaseinstaller;
 
-import java.io.IOException;
-import liquibase.Contexts;
-import liquibase.Liquibase;
-import liquibase.database.Database;
-import liquibase.database.DatabaseFactory;
-import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.LiquibaseException;
 import liquibase.logging.LogLevel;
-import liquibase.resource.ClassLoaderResourceAccessor;
-import org.apache.commons.dbcp2.BasicDataSource;
 import org.kohsuke.args4j.CmdLineException;
 import org.kohsuke.args4j.CmdLineParser;
 import org.kohsuke.args4j.ExampleMode;
@@ -33,10 +25,7 @@ import org.springframework.beans.factory.support.RootBeanDefinition;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 
-import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * Created by gibsodom on 08/12/2015.
@@ -62,8 +51,6 @@ public class Application
 
     @Option(name = "-db.name", usage = "Specifies the name of the database to be created or updated.")
     private String dbName;
-
-    private String fullConnectionString;
 
     private DatabaseProperties properties = loadProperties(DatabaseProperties.class);
 
@@ -95,7 +82,7 @@ public class Application
             return;
         }
         checkArgs();
-        applySchema();
+        new DatabaseConnector().applySchema(connectionString, username, password, dbName, allowDBDeletion);
     }
 
     private void checkArgs()
@@ -103,92 +90,12 @@ public class Application
         if (properties != null) {
             dbName = dbName != null ? dbName : properties.getDBName();
             connectionString = connectionString != null ? connectionString : properties.getConnectionString();
-            fullConnectionString = joinDBConnection(connectionString, dbName);
             username = username != null ? username : properties.getUser();
             password = password != null ? password : properties.getPass();
         } else if (connectionString == null || username == null || password == null) {
             throw new RuntimeException(
                 "If no properties specified, either supply a properties file with -DpropertySource or pass the arguments with"
                 + " -Ddb.connection, -Ddb.user, -Ddb.pass");
-        }
-    }
-
-    private void applySchema() throws SQLException, LiquibaseException
-    {
-
-        boolean dbExists = checkDBExists();
-
-        if (dbExists && allowDBDeletion) {
-            System.out.println();
-            System.out.println("DB - Exists, and force deletion has been specified for: " + dbName);
-            System.out.println();
-
-            BasicDataSource basicDataSourceNoDB = new BasicDataSource();
-            basicDataSourceNoDB.setUrl(connectionString);
-            basicDataSourceNoDB.setUsername(username);
-            basicDataSourceNoDB.setPassword(password);
-
-            try (Connection c = basicDataSourceNoDB.getConnection()) {
-                java.sql.Statement statement = c.createStatement();
-                statement.executeUpdate("DROP DATABASE " + dbName);
-                System.out.println("DELETED database: " + dbName);
-                dbExists = false;
-            }
-        }
-        if (!dbExists) {
-            System.out.println("about to perform DB installation from scratch.");
-
-            BasicDataSource basicDataSourceNoDB = new BasicDataSource();
-            basicDataSourceNoDB.setUrl(connectionString);
-            basicDataSourceNoDB.setUsername(username);
-            basicDataSourceNoDB.setPassword(password);
-
-            try (java.sql.Connection c = basicDataSourceNoDB.getConnection()) {
-                java.sql.Statement statement = c.createStatement();
-                statement.executeUpdate("CREATE DATABASE " + dbName);
-                System.out.println("Created new database: " + dbName);
-            }
-        }
-        updateDB();
-    }
-
-    /**
-     * Checks connection, retrieves appropriate changelog and performs database update.
-     *
-     * @throws SQLException
-     * @throws LiquibaseException
-     */
-    private void updateDB() throws SQLException, LiquibaseException
-    {
-        System.out.println("About to perform DB update.");
-        try (BasicDataSource dataSource = new BasicDataSource()) {
-            dataSource.setUrl(fullConnectionString);
-            dataSource.setUsername(username);
-            dataSource.setPassword(password);
-            try (java.sql.Connection c = dataSource.getConnection()) {
-                Database database = DatabaseFactory.getInstance().findCorrectDatabaseImplementation(new JdbcConnection(c));
-
-                // Check that the Database does indeed exist before we try to run the liquibase update.
-                Liquibase liquibase = null;
-                ClassLoaderResourceAccessor accessor = new ClassLoaderResourceAccessor();
-                try {
-                    if (accessor.getResourcesAsStream("changelog-master.xml") != null) {
-                        liquibase = new Liquibase("changelog-master.xml", new ClassLoaderResourceAccessor(), database);
-                    } else if (accessor.getResourcesAsStream("changelog.xml") != null) {
-                        liquibase = new Liquibase("changelog.xml", new ClassLoaderResourceAccessor(), database);
-                    } else {
-                        String errorMessage = "No liquibase changelog-master.xml or changelog.xml could be located";
-                        Logger.getLogger(Application.class.getName()).log(Level.SEVERE, errorMessage, this);
-                        throw new RuntimeException(errorMessage);
-                    }
-                } catch (final IOException ioe) {
-                    Logger.getLogger(Application.class.getName()).log(Level.SEVERE, ioe.getMessage(), ioe);
-                }
-
-                liquibase.getLog().setLogLevel(logLevel);
-                liquibase.update(new Contexts());
-                System.out.println("DB update finished.");
-            }
         }
     }
 
@@ -203,38 +110,4 @@ public class Application
         return propertiesApplicationContext.getBean(propertiesClass);
     }
 
-    private boolean checkDBExists() throws SQLException
-    {
-        try (BasicDataSource dataSource = new BasicDataSource()) {
-            dataSource.setUrl(fullConnectionString);
-            dataSource.setUsername(username);
-            dataSource.setPassword(password);
-            try (Connection c = dataSource.getConnection()) {
-                return true;
-            } catch (final Exception e) {
-                return false;
-            }
-        }
-    }
-
-    private String joinDBConnection(final String connectionString, final String dbName)
-    {
-        if (connectionString != null && dbName != null) {
-            if (connectionString.endsWith("/") && !dbName.startsWith("/")) {
-                return connectionString + dbName;
-            } else if (!connectionString.endsWith("/") && dbName.startsWith("/")) {
-                //Connection string must end with a /
-                this.connectionString = connectionString + "/";
-                return connectionString + dbName;
-            } else if (!connectionString.endsWith("/") && !dbName.startsWith("/")) {
-                //Connection string must end with a /
-                this.connectionString = connectionString + "/";
-                return connectionString + "/" + dbName;
-            } else if (connectionString.endsWith("/") && dbName.startsWith("/")) {
-                int index = connectionString.lastIndexOf("/");
-                return connectionString.substring(0, index) + dbName;
-            }
-        }
-        throw new RuntimeException("Must specify both db.connection and db.name");
-    }
 }
